Improved timer

+++ include/util/timer.hpp
@@ -17,7 +17,7 @@ public:
     virtual ~Timer();
     
     void start();
-    double deltaTime(); 
+    float deltaTime(); 
 };
 
 } // namespace util
